refactor: modularize clienthandler, step 1 - Breaks up the 1755 loc client_handler.rs into several modules. - Improves naming of various methods and objects. - Temporarily removes handling of Balance/Transfer. - Temporarily removes `refund` feature. - NB 1: referencing unmerged changes in safe-nd. - NB 2: unresolved compilation issues.
