+++ src/data_handler/adata_handler.rs
@@ -65,14 +65,12 @@ impl ADataHandler {
                 .put(&data)
                 .map_err(|error| error.to_string().into())
         };
-        let refund = utils::get_refund_for_put(&result);
         Some(Action::RespondToClientHandlers {
             sender: *data.name(),
             rpc: Rpc::Response {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                refund,
             },
         })
     }
@@ -110,8 +108,6 @@ impl ADataHandler {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                // Deletion is free so no refund
-                refund: None,
             },
         })
     }
@@ -130,7 +126,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetAData(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -152,7 +147,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataShell(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -174,7 +168,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataRange(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -195,7 +188,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataIndices(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -216,7 +208,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataLastEntry(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -243,7 +234,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataOwners(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -266,7 +256,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetPubADataUserPermissions(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -289,7 +278,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetUnpubADataUserPermissions(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -322,7 +310,6 @@ impl ADataHandler {
                 requester,
                 response,
                 message_id,
-                refund: None,
             },
         })
     }
@@ -344,7 +331,6 @@ impl ADataHandler {
                 requester,
                 response: Response::GetADataValue(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -530,14 +516,12 @@ impl ADataHandler {
                     .put(&adata)
                     .map_err(|error| error.to_string().into())
             });
-        let refund = utils::get_refund_for_put(&result);
         Some(Action::RespondToClientHandlers {
             sender: *address.name(),
             rpc: Rpc::Response {
                 requester: requester.clone(),
                 response: Response::Mutation(result),
                 message_id,
-                refund,
             },
         })
     }
