+++ src/data_handler/idata_handler.rs
@@ -64,14 +64,12 @@ impl IDataHandler {
 
         let client_id = requester.clone();
         let respond = |result: NdResult<()>| {
-            let refund = utils::get_refund_for_put(&result);
             Some(Action::RespondToClientHandlers {
                 sender: data_name,
                 rpc: Rpc::Response {
                     requester: client_id,
                     response: Response::Mutation(result),
                     message_id,
-                    refund,
                 },
             })
         };
@@ -136,7 +134,6 @@ impl IDataHandler {
                     requester: client_id,
                     response: Response::Mutation(result),
                     message_id,
-                    refund: None,
                 },
             })
         };
@@ -182,7 +179,6 @@ impl IDataHandler {
                     requester: client_id,
                     response: Response::GetIData(result),
                     message_id,
-                    refund: None,
                 },
             })
         };
@@ -281,7 +277,6 @@ impl IDataHandler {
                     requester: idata_op.client().clone(),
                     response: Response::Mutation(Ok(())),
                     message_id,
-                    refund: None,
                 },
             })
     }
@@ -348,9 +343,6 @@ impl IDataHandler {
                         requester: idata_op.client().clone(),
                         response: Response::Mutation(response),
                         message_id,
-                        // Deleting data is free so, no refund
-                        // This field can be put to use when deletion is incentivised
-                        refund: None,
                     },
                 }
             })
