+++ src/data_handler/idata_holder.rs
@@ -60,14 +60,12 @@ impl IDataHolder {
                 .put(&data)
                 .map_err(|error| error.to_string().into())
         };
-        let refund = utils::get_refund_for_put(&result);
         Some(Action::RespondToOurDataHandlers {
             sender: *self.id.name(),
             rpc: Rpc::Response {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                refund,
             },
         })
     }
@@ -100,7 +98,6 @@ impl IDataHolder {
                 requester: client.clone(),
                 response: Response::GetIData(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -146,7 +143,6 @@ impl IDataHolder {
                 requester: client.clone(),
                 response: Response::Mutation(result),
                 message_id,
-                refund: None,
             },
         })
     }
