+++ src/data_handler/idata_op.rs
@@ -175,7 +175,6 @@ impl IDataOp {
                     requester: self.client().clone(),
                     response,
                     message_id,
-                    refund: None,
                 },
             })
         }
