+++ src/data_handler/mdata_handler.rs
@@ -106,14 +106,12 @@ impl MDataHandler {
                     .put(&mdata)
                     .map_err(|error| error.to_string().into())
             });
-        let refund = utils::get_refund_for_put(&result);
         Some(Action::RespondToClientHandlers {
             sender: *address.name(),
             rpc: Rpc::Response {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                refund,
             },
         })
     }
@@ -132,14 +130,12 @@ impl MDataHandler {
                 .put(&data)
                 .map_err(|error| error.to_string().into())
         };
-        let refund = utils::get_refund_for_put(&result);
         Some(Action::RespondToClientHandlers {
             sender: *data.name(),
             rpc: Rpc::Response {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                refund,
             },
         })
     }
@@ -173,8 +169,6 @@ impl MDataHandler {
                 requester,
                 response: Response::Mutation(result),
                 message_id,
-                // Deletion is free so no refund
-                refund: None,
             },
         })
     }
@@ -247,7 +241,6 @@ impl MDataHandler {
                 requester,
                 response: Response::GetMData(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -269,7 +262,6 @@ impl MDataHandler {
                 requester,
                 response: Response::GetMDataShell(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -291,7 +283,6 @@ impl MDataHandler {
                 requester,
                 response: Response::GetMDataVersion(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -327,7 +318,6 @@ impl MDataHandler {
                 requester,
                 response,
                 message_id,
-                refund: None,
             },
         })
     }
@@ -349,7 +339,6 @@ impl MDataHandler {
                 requester,
                 response: Response::ListMDataKeys(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -374,7 +363,6 @@ impl MDataHandler {
                 requester,
                 response,
                 message_id,
-                refund: None,
             },
         })
     }
@@ -399,7 +387,6 @@ impl MDataHandler {
                 requester,
                 response,
                 message_id,
-                refund: None,
             },
         })
     }
@@ -421,7 +408,6 @@ impl MDataHandler {
                 requester,
                 response: Response::ListMDataPermissions(result),
                 message_id,
-                refund: None,
             },
         })
     }
@@ -444,7 +430,6 @@ impl MDataHandler {
                 requester,
                 response: Response::ListMDataUserPermissions(result),
                 message_id,
-                refund: None,
             },
         })
     }
