+++ src/utils.rs
@@ -6,15 +6,12 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::client_handler::COST_OF_PUT;
 use crate::{rpc::Rpc, vault::Init, Result};
 use bincode;
 use log::{error, trace};
 use pickledb::{PickleDb, PickleDbDumpPolicy};
 use rand::{distributions::Standard, CryptoRng, Rng};
-use safe_nd::{
-    ClientPublicId, Coins, IDataAddress, PublicId, PublicKey, Request, Result as NdResult, XorName,
-};
+use safe_nd::{ClientPublicId, IDataAddress, PublicId, PublicKey, Request, XorName};
 use serde::Serialize;
 use std::{borrow::Cow, fs, path::Path};
 use unwrap::unwrap;
@@ -226,10 +223,10 @@ pub(crate) fn authorisation_kind(request: &Request) -> AuthorisationKind {
     }
 }
 
-pub(crate) fn get_refund_for_put<T>(result: &NdResult<T>) -> Option<Coins> {
-    if result.is_err() {
-        Some(COST_OF_PUT)
-    } else {
-        None
-    }
-}
+// pub(crate) fn get_refund_for_put<T>(result: &NdResult<T>) -> Option<Coins> {
+//     if result.is_err() {
+//         Some(COST_OF_PUT)
+//     } else {
+//         None
+//     }
+// }
